+++ modules/Food_Service/User.inc.php
@@ -5,20 +5,16 @@ require_once 'ProgramFunctions/TipMessage.fnc.php';
 if ( $_REQUEST['modfunc'] === 'update' )
 {
 	if ( UserStaffID()
-		&& AllowEdit() )
+		&& AllowEdit()
+		&& ! empty( $_REQUEST['food_service'] ) )
 	{
-		if ( ! empty( $_REQUEST['food_service'] ) )
-		{
-			$sql = "UPDATE food_service_staff_accounts SET ";
+		DBUpdate(
+			'food_service_staff_accounts',
+			array_map( 'trim', $_REQUEST['food_service'] ),
+			[ 'STAFF_ID' => UserStaffID() ]
+		);
 
-			foreach ( (array) $_REQUEST['food_service'] as $column_name => $value )
-			{
-				$sql .= DBEscapeIdentifier( $column_name ) . "='" . trim( $value ) . "',";
-			}
-
-			$sql = mb_substr( $sql, 0, -1 ) . " WHERE STAFF_ID='" . UserStaffID() . "'";
-			DBQuery( $sql );
-		}
+		$note[] = button( 'check' ) . ' ' . _( 'Your changes were saved.' );
 	}
 
 	// $_REQUEST['modfunc'] = false;